feat: abstract util and remove TODO
