+++ src/features/defi/providers/thorchain-savers/components/ThorchainSaversManager/Deposit/components/Deposit.tsx
@@ -47,11 +47,8 @@ import {
   getErc20Allowance,
   getFees,
 } from 'lib/utils/evm'
-import type { GetThorchainSaversDepositQuoteQueryKey } from 'lib/utils/thorchain/hooks/useGetThorchainSaversDepositQuoteQuery'
-import {
-  queryFn as getThorchainSaversDepositQuoteQueryFn,
-  useGetThorchainSaversDepositQuoteQuery,
-} from 'lib/utils/thorchain/hooks/useGetThorchainSaversDepositQuoteQuery'
+import { fetchHasEnoughBalanceForTxPlusFeesPlusSweep } from 'lib/utils/thorchain'
+import { useGetThorchainSaversDepositQuoteQuery } from 'lib/utils/thorchain/hooks/useGetThorchainSaversDepositQuoteQuery'
 import type { EstimatedFeesQueryKey } from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
 import {
   queryFn as getEstimatedFeesQueryFn,
@@ -450,31 +447,6 @@ export const Deposit: React.FC<DepositProps> = ({
   const { data: isSweepNeeded, isLoading: isSweepNeededLoading } =
     useIsSweepNeededQuery(isSweepNeededArgs)
 
-  const getHasEnoughBalanceForTxPlusFeesPlusSweep = useCallback(
-    ({
-      balanceCryptoBaseUnit,
-      amountCryptoPrecision,
-      txFeeCryptoBaseUnit,
-      precision,
-      sweepTxFeeCryptoBaseUnit,
-    }: {
-      balanceCryptoBaseUnit: string
-      amountCryptoPrecision: string
-      txFeeCryptoBaseUnit: string
-      precision: number
-      sweepTxFeeCryptoBaseUnit: string
-    }) => {
-      const balanceCryptoBaseUnitBn = bnOrZero(balanceCryptoBaseUnit)
-      if (balanceCryptoBaseUnitBn.isZero()) return false
-
-      return bnOrZero(amountCryptoPrecision)
-        .plus(fromBaseUnit(txFeeCryptoBaseUnit, precision ?? 0))
-        .plus(fromBaseUnit(sweepTxFeeCryptoBaseUnit, precision ?? 0))
-        .lte(fromBaseUnit(balanceCryptoBaseUnitBn, precision ?? 0))
-    },
-    [],
-  )
-
   const handleContinue = useCallback(
     async (formValues: DepositValues) => {
       if (
@@ -593,132 +565,10 @@ export const Deposit: React.FC<DepositProps> = ({
     browserHistory.goBack()
   }, [browserHistory])
 
-  // TODO(gomes): abstract me into a pure method when implementing sweep for withdraw, and handle non-UTXO chains
-  const fetchHasEnoughBalanceForTxPlusFeesPlusSweep = useCallback(
-    async (valueCryptoPrecision: string) => {
-      const isUtxoChain = isUtxoChainId(asset.chainId)
-      const estimateFeesQueryEnabled = Boolean(fromAddress && accountId && isUtxoChain)
-
-      const amountCryptoBaseUnit = toBaseUnit(valueCryptoPrecision, asset.precision)
-
-      const thorchainSaversDepositQuoteQueryKey: GetThorchainSaversDepositQuoteQueryKey = [
-        'thorchainSaversDepositQuote',
-        { asset, amountCryptoBaseUnit },
-      ]
-
-      const _thorchainSaversDepositQuote = await queryClient.fetchQuery({
-        queryKey: thorchainSaversDepositQuoteQueryKey,
-        queryFn: getThorchainSaversDepositQuoteQueryFn,
-        staleTime: 5000,
-      })
-
-      const estimatedFeesQueryArgs = {
-        estimateFeesInput: {
-          cryptoAmount: valueCryptoPrecision,
-          assetId,
-          to: _thorchainSaversDepositQuote?.inbound_address ?? '',
-          sendMax: false,
-          accountId: accountId ?? '',
-          contractAddress: undefined,
-        },
-        asset,
-        assetMarketData: marketData,
-        enabled: estimateFeesQueryEnabled,
-      }
-
-      const estimatedFeesQueryKey: EstimatedFeesQueryKey = ['estimateFees', estimatedFeesQueryArgs]
-
-      const _estimatedFeesData = estimateFeesQueryEnabled
-        ? await queryClient.fetchQuery({
-            queryKey: estimatedFeesQueryKey,
-            queryFn: getEstimatedFeesQueryFn,
-          })
-        : undefined
-
-      const _hasEnoughBalanceForTxPlusFees = getHasEnoughBalanceForTxPlusFees({
-        precision: asset.precision,
-        balanceCryptoBaseUnit,
-        amountCryptoPrecision: valueCryptoPrecision,
-        txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '',
-      })
-
-      const isSweepNeededQueryEnabled = Boolean(
-        bnOrZero(valueCryptoPrecision).gt(0) &&
-          _estimatedFeesData &&
-          _hasEnoughBalanceForTxPlusFees,
-      )
-
-      const isSweepNeededQueryArgs = {
-        assetId,
-        address: fromAddress,
-        txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '0',
-        amountCryptoBaseUnit,
-      }
-
-      const isSweepNeededQueryKey: IsSweepNeededQueryKey = ['isSweepNeeded', isSweepNeededQueryArgs]
-
-      const _isSweepNeeded = isSweepNeededQueryEnabled
-        ? await queryClient.fetchQuery({
-            queryKey: isSweepNeededQueryKey,
-            queryFn: isSweepNeededQueryFn,
-          })
-        : undefined
-
-      const isEstimateSweepFeesQueryEnabled = Boolean(_isSweepNeeded && accountId && isUtxoChain)
-
-      const estimatedSweepFeesQueryArgs = {
-        asset,
-        assetMarketData: marketData,
-        estimateFeesInput: {
-          cryptoAmount: '0',
-          assetId,
-          to: fromAddress ?? '',
-          sendMax: true,
-          accountId: accountId ?? '',
-          contractAddress: undefined,
-        },
-        enabled: isEstimateSweepFeesQueryEnabled,
-      }
-
-      const estimatedSweepFeesQueryKey: EstimatedFeesQueryKey = [
-        'estimateFees',
-        estimatedSweepFeesQueryArgs,
-      ]
-
-      const _estimatedSweepFeesData = isEstimateSweepFeesQueryEnabled
-        ? await queryClient.fetchQuery({
-            queryKey: estimatedSweepFeesQueryKey,
-            queryFn: getEstimatedFeesQueryFn,
-          })
-        : undefined
-
-      const hasEnoughBalanceForTxPlusFeesPlusSweep =
-        bnOrZero(balanceCryptoBaseUnit).gt(0) &&
-        getHasEnoughBalanceForTxPlusFeesPlusSweep({
-          precision: asset.precision,
-          balanceCryptoBaseUnit,
-          amountCryptoPrecision: valueCryptoPrecision,
-          txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '0',
-          sweepTxFeeCryptoBaseUnit: _estimatedSweepFeesData?.txFeeCryptoBaseUnit ?? '0',
-        })
-
-      return hasEnoughBalanceForTxPlusFeesPlusSweep
-    },
-    [
-      accountId,
-      asset,
-      assetId,
-      balanceCryptoBaseUnit,
-      fromAddress,
-      getHasEnoughBalanceForTxPlusFees,
-      getHasEnoughBalanceForTxPlusFeesPlusSweep,
-      marketData,
-      queryClient,
-    ],
-  )
-
   const validateCryptoAmount = useCallback(
     async (value: string) => {
+      if (!accountId) return
+
       const valueCryptoBaseUnit = toBaseUnit(value, asset.precision)
       const balanceCryptoPrecision = bn(fromBaseUnit(balanceCryptoBaseUnit, asset.precision))
 
@@ -746,23 +596,30 @@ export const Deposit: React.FC<DepositProps> = ({
         return translate('trade.errors.amountTooSmall', { minLimit: outboundFeeLimit })
 
       const hasEnoughBalanceForTxPlusFeesPlusSweep =
-        await fetchHasEnoughBalanceForTxPlusFeesPlusSweep(value)
+        await fetchHasEnoughBalanceForTxPlusFeesPlusSweep({
+          amountCryptoPrecision: value,
+          accountId,
+          asset,
+          type: 'deposit',
+          fromAddress,
+        })
 
       return hasEnoughBalanceForTxPlusFeesPlusSweep || 'common.insufficientFunds'
     },
     [
-      asset.precision,
-      asset.symbol,
+      asset,
       balanceCryptoBaseUnit,
       assetId,
       outboundFeeCryptoBaseUnit,
       translate,
-      fetchHasEnoughBalanceForTxPlusFeesPlusSweep,
+      accountId,
+      fromAddress,
     ],
   )
 
   const validateFiatAmount = useCallback(
     async (value: string) => {
+      if (!accountId) return
       const valueCryptoPrecision = bnOrZero(value).div(marketData.price)
       const balanceCryptoPrecision = bn(fromBaseUnit(balanceCryptoBaseUnit, asset.precision))
 
@@ -789,18 +646,24 @@ export const Deposit: React.FC<DepositProps> = ({
         return translate('trade.errors.amountTooSmall', { minLimit: outboundFeeLimit })
 
       const hasEnoughBalanceForTxPlusFeesPlusSweep =
-        await fetchHasEnoughBalanceForTxPlusFeesPlusSweep(valueCryptoPrecision.toFixed())
+        await fetchHasEnoughBalanceForTxPlusFeesPlusSweep({
+          amountCryptoPrecision: valueCryptoPrecision.toFixed(),
+          accountId,
+          asset,
+          type: 'deposit',
+          fromAddress,
+        })
       return hasEnoughBalanceForTxPlusFeesPlusSweep || 'common.insufficientFunds'
     },
     [
       marketData.price,
       balanceCryptoBaseUnit,
-      asset.precision,
-      asset.symbol,
+      asset,
       assetId,
       outboundFeeCryptoBaseUnit,
       translate,
-      fetchHasEnoughBalanceForTxPlusFeesPlusSweep,
+      accountId,
+      fromAddress,
     ],
   )
